Refactor LR automaton builder into modules
