+++ glr-core/src/lib.rs
@@ -68,6 +68,9 @@ pub use GLRError as GlrError;
 /// Conflict inspection API for analyzing GLR parse table conflicts
 pub mod conflict_inspection;
 
+mod automaton;
+pub use automaton::build_lr1_automaton;
+
 // Re-export key types from adze-ir for API consumers
 /// Re-exported IR types used throughout GLR construction.
 pub use adze_ir::{Grammar, RuleId, StateId, SymbolId};
@@ -171,261 +174,6 @@ pub use precedence_compare::{
 pub use symbol_comparison::{compare_symbols, compare_versions_with_symbols};
 #[doc(hidden)]
 pub use version_info::{CompareResult, VersionInfo, compare_versions};
-
-// ============================================================================
-// EOF Symbol Sentinel Handling
-// ============================================================================
-//
-// FirstFollowSets uses SymbolId(0) as an internal sentinel to represent EOF
-// in FIRST/FOLLOW set computations. However, the actual EOF symbol in the
-// parse table is dynamically assigned to avoid conflicts with grammar symbols.
-//
-// Use these helpers at the boundary where FIRST/FOLLOW results become real
-// lookahead symbols in the parse table.
-
-/// Internal EOF sentinel used by FirstFollowSets.
-/// This is NOT the actual EOF symbol - use `parse_table.eof_symbol` for that.
-const EOF_SENTINEL: SymbolId = SymbolId(0);
-
-/// Map a symbol from FOLLOW set output to actual parse table symbol.
-/// Replaces the EOF sentinel (SymbolId(0)) with the actual EOF symbol.
-#[inline]
-fn map_follow_symbol(sym: SymbolId, eof_symbol: SymbolId) -> SymbolId {
-    if sym == EOF_SENTINEL { eof_symbol } else { sym }
-}
-
-// Precedence resolution structures
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-enum Assoc {
-    Left,
-    Right,
-    None,
-}
-
-#[derive(Copy, Clone, Debug)]
-struct TokPrec {
-    prec: u8,
-    assoc: Assoc,
-}
-
-#[derive(Copy, Clone, Debug)]
-struct RulePrec {
-    prec: u8,
-    assoc: Assoc,
-}
-
-struct PrecTables {
-    // table-indexed; entries 0..token_count-1 may be Some(..); others None
-    tok_prec_by_index: Vec<Option<TokPrec>>,
-    // production_id -> precedence and associativity
-    rule_prec: Vec<RulePrec>,
-}
-
-fn build_prec_tables(
-    grammar: &Grammar,
-    symbol_to_index: &BTreeMap<SymbolId, usize>,
-    token_count: u32,
-    production_count: u32,
-) -> PrecTables {
-    use adze_ir::{Associativity, PrecedenceKind};
-
-    // Guard rail: ensure we have the right table structure
-    // Note: token_count can be 0 for empty grammars (only EOF token)
-    debug_assert!(production_count > 0, "production_count must be positive");
-
-    // token precedence by table index
-    let mut tok_prec_by_index = vec![None; symbol_to_index.len()];
-    let tok_prec_len = tok_prec_by_index.len(); // Capture length before closure
-
-    // Helper: set token precedence preferring higher numeric level
-    let mut set_tok_prec = |tok_idx: usize, new: TokPrec| {
-        if tok_idx >= tok_prec_by_index.len() {
-            return; // Guard against out-of-bounds
-        }
-        tok_prec_by_index[tok_idx] = match tok_prec_by_index[tok_idx] {
-            None => Some(new),
-            Some(old) => Some(if new.prec > old.prec { new } else { old }),
-        };
-    };
-
-    // production precedence: explicit if present, else rightmost-terminal precedence
-    let mut rule_prec = vec![
-        RulePrec {
-            prec: 0,
-            assoc: Assoc::None
-        };
-        production_count as usize
-    ];
-
-    // Two-pass approach: first collect rule precedence, then derive token precedence
-    for rules in grammar.rules.values() {
-        for rule in rules {
-            let pid = rule.production_id.0 as usize;
-            // Skip invalid production IDs (e.g., u16::MAX used as sentinel)
-            if pid >= production_count as usize {
-                continue;
-            }
-
-            // 1) Rule precedence (explicit)
-            let explicit = rule.precedence.and_then(|p| {
-                if let PrecedenceKind::Static(level) = p {
-                    Some(level as u8)
-                } else {
-                    None
-                }
-            });
-
-            // Get rule associativity (defaults to None if not specified)
-            let rule_assoc = rule
-                .associativity
-                .map(|assoc| match assoc {
-                    Associativity::Left => Assoc::Left,
-                    Associativity::Right => Assoc::Right,
-                    Associativity::None => Assoc::None,
-                })
-                .unwrap_or(Assoc::None);
-
-            // 2) Derive token precedence from the rightmost terminal if this rule carries
-            //    an explicit precedence (+ associativity) attribute
-            if let Some(level) = explicit {
-                // Find rightmost terminal in RHS
-                let tok_idx_opt = rule.rhs.iter().rev().find_map(|sym| {
-                    if let Symbol::Terminal(id) = sym {
-                        symbol_to_index.get(id).copied()
-                    } else {
-                        None
-                    }
-                });
-
-                if let Some(tok_idx) = tok_idx_opt
-                    && tok_idx < tok_prec_len
-                {
-                    set_tok_prec(
-                        tok_idx,
-                        TokPrec {
-                            prec: level,
-                            assoc: rule_assoc,
-                        },
-                    );
-                }
-            }
-
-            // 3) Store the rule precedence AND associativity
-            rule_prec[pid] = RulePrec {
-                prec: explicit.unwrap_or(0),
-                assoc: rule_assoc,
-            };
-        }
-    }
-
-    // Second pass: for rules without explicit precedence, inherit from rightmost terminal
-    for rules in grammar.rules.values() {
-        for rule in rules {
-            let pid = rule.production_id.0 as usize;
-            if pid >= production_count as usize {
-                continue;
-            }
-
-            // Skip if already has precedence
-            if rule_prec[pid].prec > 0 {
-                continue;
-            }
-
-            // Inherit from rightmost terminal token precedence AND associativity
-            let derived = rule
-                .rhs
-                .iter()
-                .rev()
-                .find_map(|sym| {
-                    if let Symbol::Terminal(id) = sym {
-                        symbol_to_index.get(id).and_then(|&idx| {
-                            if (idx as u32) < token_count {
-                                tok_prec_by_index[idx]
-                            } else {
-                                None
-                            }
-                        })
-                    } else {
-                        None
-                    }
-                })
-                .unwrap_or(TokPrec {
-                    prec: 0,
-                    assoc: Assoc::None,
-                });
-
-            rule_prec[pid] = RulePrec {
-                prec: derived.prec,
-                assoc: derived.assoc,
-            };
-        }
-    }
-
-    PrecTables {
-        tok_prec_by_index,
-        rule_prec,
-    }
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-enum PrecDecision {
-    PreferShift,
-    PreferReduce,
-    Error,
-    NoInfo,
-}
-
-#[inline]
-fn decide_with_precedence(
-    lookahead_tok_idx: usize, // table-index of token
-    reduce_prod_id: u16,      // production id from Action::Reduce
-    prec: &PrecTables,
-) -> PrecDecision {
-    // Guard rail: production ID must be valid
-    if reduce_prod_id as usize >= prec.rule_prec.len() {
-        return PrecDecision::NoInfo;
-    }
-
-    let tokp = match prec
-        .tok_prec_by_index
-        .get(lookahead_tok_idx)
-        .and_then(|o| *o)
-    {
-        Some(p) => p,
-        None => return PrecDecision::NoInfo,
-    };
-    let rulep = prec.rule_prec[reduce_prod_id as usize];
-
-    // If either has no precedence info (0), can't decide
-    if tokp.prec == 0 || rulep.prec == 0 {
-        return PrecDecision::NoInfo;
-    }
-
-    use core::cmp::Ordering::*;
-    // FIX: Use RULE's associativity, not token's!
-    // When precedences are equal, the rule's associativity determines shift vs reduce
-    match (tokp.prec.cmp(&rulep.prec), rulep.assoc) {
-        (Greater, _) => PrecDecision::PreferShift,
-        (Less, _) => PrecDecision::PreferReduce,
-        (Equal, Assoc::Left) => PrecDecision::PreferReduce, // left-assoc: reduce first
-        (Equal, Assoc::Right) => PrecDecision::PreferShift, // right-assoc: shift first
-        (Equal, Assoc::None) => PrecDecision::Error,
-    }
-}
-
-// Handle reduce/reduce conflicts (prefer higher rule precedence, tie -> lowest pid)
-#[inline]
-fn decide_reduce_reduce(a: u16, b: u16, prec: &PrecTables) -> u16 {
-    let pa = prec.rule_prec.get(a as usize).map(|r| r.prec).unwrap_or(0);
-    let pb = prec.rule_prec.get(b as usize).map(|r| r.prec).unwrap_or(0);
-    if pa > pb {
-        a
-    } else if pb > pa {
-        b
-    } else {
-        a.min(b)
-    }
-}
 
 // Public API exports
 /// The main GLR parser driver.
@@ -2472,52 +2220,6 @@ fn can_derive_start(grammar: &Grammar, symbol: SymbolId, start: SymbolId) -> boo
     false
 }
 
-/// Normalize action cells for deterministic output
-/// Sorts actions by type and value, and removes duplicates
-fn normalize_action_table(action_table: &mut Vec<Vec<ActionCell>>) {
-    for row in action_table.iter_mut() {
-        for cell in row.iter_mut() {
-            normalize_action_cell(cell);
-        }
-    }
-}
-
-/// Normalize a single action cell (recursive for Fork actions)
-fn normalize_action_cell(cell: &mut ActionCell) {
-    for action in cell.iter_mut() {
-        normalize_action(action);
-    }
-    cell.sort_by_key(action_sort_key);
-    cell.dedup();
-}
-
-/// Normalize a single action (recursively normalizes Fork contents)
-fn normalize_action(action: &mut Action) {
-    if let Action::Fork(inner) = action {
-        for inner_action in inner.iter_mut() {
-            normalize_action(inner_action);
-        }
-        inner.sort_by_key(action_sort_key);
-        inner.dedup();
-    }
-}
-
-/// Generate a sort key for actions to ensure deterministic ordering
-fn action_sort_key(action: &Action) -> (u8, u16, u16, u16) {
-    match action {
-        Action::Shift(s) => (0, s.0, 0, 0),
-        Action::Reduce(r) => (1, r.0, 0, 0),
-        Action::Accept => (2, 0, 0, 0),
-        Action::Error => (3, 0, 0, 0),
-        Action::Recover => (4, 0, 0, 0),
-        Action::Fork(inner) => {
-            // Tie-break forks by length (minor) and first key (major)
-            let first = inner.first().map(action_sort_key).unwrap_or((0, 0, 0, 0));
-            (5, first.1, first.2, inner.len() as u16)
-        }
-    }
-}
-
 /// Build LR(1) automaton (parse table) from grammar.
 ///
 /// Constructs an augmented grammar, builds the canonical LR(1) collection,
@@ -2551,723 +2253,6 @@ fn action_sort_key(action: &Action) -> (u8, u16, u16, u16) {
 /// });
 /// assert!(has_accept, "table must have an Accept action");
 /// ```
-pub fn build_lr1_automaton(
-    grammar: &Grammar,
-    first_follow: &FirstFollowSets,
-) -> Result<ParseTable, GLRError> {
-    // Debug: Print some rules to see their structure
-    let mut rule_count = 0;
-    for rule in grammar.all_rules() {
-        if rule_count >= 10 {
-            break;
-        }
-        let mut rhs_str = String::new();
-        for sym in &rule.rhs {
-            match sym {
-                Symbol::Terminal(id) => rhs_str.push_str(&format!("T({}) ", id.0)),
-                Symbol::NonTerminal(id) => rhs_str.push_str(&format!("NT({}) ", id.0)),
-                _ => rhs_str.push_str("? "),
-            }
-        }
-        rule_count += 1;
-    }
-
-    // Build stable symbol partitions for table columns:
-    // EOF, internal terminals, external terminals, then nonterminals.
-    // Internal terminals are not limited to grammar.tokens; they may also come
-    // from literal/punctuation terminals that only appear in rule RHS.
-    let nonterminal_symbols: BTreeSet<SymbolId> = grammar.rules.keys().copied().collect();
-    let external_symbols: BTreeSet<SymbolId> =
-        grammar.externals.iter().map(|e| e.symbol_id).collect();
-    let mut rhs_terminals: BTreeSet<SymbolId> = BTreeSet::new();
-    for rule in grammar.all_rules() {
-        for sym in &rule.rhs {
-            if let Symbol::Terminal(id) = sym {
-                rhs_terminals.insert(*id);
-            }
-        }
-    }
-
-    // Calculate EOF symbol ID to avoid collisions with all known symbols,
-    // including RHS-only terminals.
-    let max_symbol = grammar
-        .tokens
-        .keys()
-        .chain(grammar.rule_names.keys())
-        .chain(nonterminal_symbols.iter())
-        .chain(external_symbols.iter())
-        .chain(rhs_terminals.iter())
-        .map(|s| s.0)
-        .max()
-        .unwrap_or(0);
-    let eof_symbol = SymbolId(max_symbol.checked_add(1).ok_or_else(|| {
-        GLRError::StateMachine("EOF symbol would overflow u16: grammar has too many symbols".into())
-    })?);
-
-    // Create augmented grammar with S' -> S $ rule
-    let mut augmented_grammar = grammar.clone();
-
-    // Find the original start symbol
-    let original_start =
-        grammar
-            .start_symbol()
-            .ok_or(GLRError::GrammarError(GrammarError::UnresolvedSymbol(
-                SymbolId(0),
-            )))?;
-
-    if let Some(_name) = grammar.rule_names.get(&original_start) {}
-
-    // Create a new start symbol S' with an ID that won't conflict with existing symbols
-    // Since eof_symbol = max_symbol + 1, we use max_symbol + 2 for augmented_start
-    let augmented_start_id = max_symbol.checked_add(2).ok_or_else(|| {
-        GLRError::StateMachine(
-            "Augmented start symbol would overflow u16: grammar has too many symbols".into(),
-        )
-    })?;
-    let augmented_start = SymbolId(augmented_start_id);
-
-    // Find the max production ID in the original grammar
-    let max_production_id = grammar
-        .all_rules()
-        .map(|r| r.production_id.0)
-        .max()
-        .unwrap_or(0);
-    let augmented_production_id = max_production_id
-        .checked_add(1)
-        .ok_or_else(|| GLRError::StateMachine("Production ID overflow".into()))?;
-
-    // Add S' -> S rule (we'll handle $ implicitly in the LR construction)
-    let augmented_rule = Rule {
-        lhs: augmented_start,
-        rhs: vec![Symbol::NonTerminal(original_start)],
-        precedence: None,
-        associativity: None,
-        fields: vec![],
-        production_id: ProductionId(augmented_production_id),
-    };
-    augmented_grammar
-        .rules
-        .insert(augmented_start, vec![augmented_rule]);
-    augmented_grammar
-        .rule_names
-        .insert(augmented_start, "$start".to_string());
-
-    // Build canonical collection of LR(1) item sets with augmented grammar
-    let collection = ItemSetCollection::build_canonical_collection_augmented(
-        &augmented_grammar,
-        first_follow,
-        augmented_start,
-        original_start,
-        eof_symbol,
-    );
-
-    // Create mapping from symbol IDs to table indices
-    let mut symbol_to_index = BTreeMap::new();
-
-    // IMPORTANT: EOF symbol must always have index 0 in Tree-sitter
-    symbol_to_index.insert(eof_symbol, 0);
-
-    // Group symbols in Tree-sitter order:
-    // 1. EOF first (index 0)
-    // 2. Regular tokens (terminals)
-    // 3. External tokens (terminals)
-    // 4. Non-terminals last (gotos)
-
-    // Internal terminals are:
-    // (declared tokens ∪ RHS terminals) − nonterminals − externals − EOF.
-    let mut internal_terminals: BTreeSet<SymbolId> = grammar.tokens.keys().copied().collect();
-    internal_terminals.extend(rhs_terminals.iter().copied());
-    internal_terminals.remove(&eof_symbol);
-    for id in &external_symbols {
-        internal_terminals.remove(id);
-    }
-    for id in &nonterminal_symbols {
-        internal_terminals.remove(id);
-    }
-
-    let mut internal_tokens: Vec<SymbolId> = internal_terminals.into_iter().collect();
-    internal_tokens.sort_by_key(|s| s.0);
-    for &id in &internal_tokens {
-        if !symbol_to_index.contains_key(&id) {
-            let idx = symbol_to_index.len();
-            symbol_to_index.insert(id, idx);
-        }
-    }
-
-    // Collect and sort external tokens
-    let mut ext_tokens: Vec<SymbolId> = external_symbols.iter().copied().collect();
-    ext_tokens.sort_by_key(|s| s.0);
-    for &id in &ext_tokens {
-        if !symbol_to_index.contains_key(&id) {
-            let idx = symbol_to_index.len();
-            symbol_to_index.insert(id, idx);
-        }
-    }
-
-    // Collect and sort non-terminals
-    let mut non_terminals: Vec<SymbolId> = nonterminal_symbols.iter().copied().collect();
-    non_terminals.sort_by_key(|s| s.0);
-    for id in non_terminals {
-        if !symbol_to_index.contains_key(&id) {
-            let idx = symbol_to_index.len();
-            symbol_to_index.insert(id, idx);
-        }
-    }
-
-    // Any remaining symbols indicate partition drift and should be investigated.
-    let mut other_symbols: Vec<SymbolId> = grammar
-        .rule_names
-        .keys()
-        .cloned()
-        .filter(|id| !symbol_to_index.contains_key(id))
-        .collect();
-    other_symbols.sort_by_key(|s| s.0);
-    if !other_symbols.is_empty() {
-        return Err(GLRError::StateMachine(format!(
-            "Unexpected symbols outside terminal/nonterminal partitions: {:?}",
-            other_symbols
-        )));
-    }
-
-    // Calculate the final symbol count after adding all symbols including EOF
-    let indexed_symbol_count = symbol_to_index.len();
-
-    // Create parse table with proper dimensions
-    let state_count = collection.sets.len();
-    let symbol_count = indexed_symbol_count; // Keep for compatibility
-
-    let mut action_table = vec![vec![Vec::new(); indexed_symbol_count]; state_count];
-    let mut goto_table = vec![vec![StateId(0); indexed_symbol_count]; state_count];
-
-    // Track conflicts as we build the table
-    let mut conflicts_by_state: BTreeMap<(usize, usize), Vec<Action>> = BTreeMap::new();
-
-    // Build rules for reduction and collect precedence info
-    let mut rules = Vec::new();
-    let mut dynamic_prec_by_rule = Vec::new();
-    let mut rule_assoc_by_rule = Vec::new();
-    let mut production_to_rule_id = BTreeMap::new();
-
-    for (rule_id, rule) in grammar.all_rules().enumerate() {
-        production_to_rule_id.insert(rule.production_id.0, rule_id as u16);
-        rules.push(ParseRule {
-            lhs: rule.lhs,
-            rhs_len: rule.rhs.len() as u16,
-        });
-
-        // Extract precedence value for this rule
-        let prec = match rule.precedence {
-            Some(adze_ir::PrecedenceKind::Static(p)) => p,
-            Some(adze_ir::PrecedenceKind::Dynamic(p)) => p,
-            None => 0,
-        };
-        dynamic_prec_by_rule.push(prec);
-
-        // Extract associativity for this rule
-        let assoc = match rule.associativity {
-            Some(adze_ir::Associativity::Left) => 1,
-            Some(adze_ir::Associativity::Right) => -1,
-            _ => 0,
-        };
-        rule_assoc_by_rule.push(assoc);
-    }
-
-    // Debug: Print goto table entries
-    debug_trace!(
-        "DEBUG: Collection goto table has {} entries",
-        collection.goto_table.len()
-    );
-    debug_trace!(
-        "DEBUG: Augmented grammar has {} tokens",
-        augmented_grammar.tokens.len()
-    );
-
-    // Debug: Print what tokens are in the augmented grammar
-    debug_trace!("=== Symbol Classification Debug ===");
-    debug_trace!(
-        "Tokens in augmented_grammar: {:?}",
-        augmented_grammar
-            .tokens
-            .keys()
-            .map(|k| k.0)
-            .collect::<Vec<_>>()
-    );
-    debug_trace!(
-        "Externals in augmented_grammar: {:?}",
-        augmented_grammar
-            .externals
-            .iter()
-            .map(|e| e.symbol_id.0)
-            .collect::<Vec<_>>()
-    );
-    debug_trace!("Original grammar tokens: {}", grammar.tokens.len());
-    debug_trace!(
-        "Collection goto_table size: {}",
-        collection.goto_table.len()
-    );
-
-    // Debug state 0 specifically
-    let state0_gotos: Vec<_> = collection
-        .goto_table
-        .iter()
-        .filter(|((from, _), _)| from.0 == 0)
-        .collect();
-    debug_trace!("State 0 has {} goto entries", state0_gotos.len());
-    for ((_, _symbol), _to_state) in &state0_gotos {
-        debug_trace!("  Symbol {} -> State {}", _symbol.0, _to_state.0);
-    }
-
-    // First, add shift actions from goto table for terminals
-    // This must be done BEFORE reduce actions to enable shift/reduce conflict detection
-    let mut _terminal_count = 0;
-    let mut _non_terminal_count = 0;
-
-    for ((from_state, symbol), to_state) in &collection.goto_table {
-        // Check if this symbol is a terminal using the tracking from collection
-        let is_terminal = collection
-            .symbol_is_terminal
-            .get(symbol)
-            .copied()
-            .unwrap_or(*symbol == eof_symbol); // EOF is a terminal
-
-        if from_state.0 == 0 {
-            debug_trace!(
-                "State 0 goto entry: symbol {} -> state {}, is_terminal={} (in tokens={}, in externals={}, is EOF={})",
-                symbol.0,
-                to_state.0,
-                is_terminal,
-                augmented_grammar.tokens.contains_key(symbol),
-                augmented_grammar
-                    .externals
-                    .iter()
-                    .any(|e| e.symbol_id == *symbol),
-                symbol.0 == 0
-            );
-        }
-
-        if is_terminal {
-            _terminal_count += 1;
-            if let Some(&symbol_idx) = symbol_to_index.get(symbol) {
-                let state_idx = from_state.0 as usize;
-                if state_idx < action_table.len() && symbol_idx < action_table[state_idx].len() {
-                    // Add as a shift action
-                    let new_action = Action::Shift(*to_state);
-                    if state_idx == 0 {
-                        debug_trace!(
-                            "DEBUG: Adding shift action to state 0: symbol {} (idx={}) -> state {}",
-                            symbol.0,
-                            symbol_idx,
-                            to_state.0
-                        );
-                    }
-                    add_action_with_conflict(
-                        &mut action_table,
-                        &mut conflicts_by_state,
-                        state_idx,
-                        symbol_idx,
-                        new_action,
-                    );
-                } else if state_idx == 0 {
-                    debug_trace!(
-                        "DEBUG: SKIPPING shift for state 0: bounds check failed - state_idx={}, symbol_idx={}, action_table.len={}, inner_len={}",
-                        state_idx,
-                        symbol_idx,
-                        action_table.len(),
-                        if state_idx < action_table.len() {
-                            action_table[state_idx].len()
-                        } else {
-                            0
-                        }
-                    );
-                }
-            } else if from_state.0 == 0 {
-                debug_trace!(
-                    "DEBUG: Terminal {} not in symbol_to_index for state 0",
-                    symbol.0
-                );
-            }
-        } else {
-            _non_terminal_count += 1;
-        }
-    }
-
-    // Handle "extras" (like comments, whitespace, and external tokens marked as extras).
-    // In every state, for every "extra" token, if there isn't already a specific
-    // action, add a self-looping SHIFT action. This allows extras to appear
-    // anywhere in the grammar without changing the parser's state.
-    for state_idx in 0..state_count {
-        for extra_symbol_id in &augmented_grammar.extras {
-            if let Some(&symbol_idx) = symbol_to_index.get(extra_symbol_id) {
-                // Check if an action already exists for this extra token in this state.
-                // Only add self-loop if no action exists yet (empty cell means no action)
-                if action_table[state_idx][symbol_idx].is_empty() {
-                    // Add a self-looping shift that stays in the same state
-                    action_table[state_idx][symbol_idx]
-                        .push(Action::Shift(StateId(state_idx as u16)));
-                }
-            }
-        }
-    }
-
-    // Now fill action table with reduce actions
-    for item_set in &collection.sets {
-        let state_idx = item_set.id.0 as usize;
-
-        for item in &item_set.items {
-            if item.is_reduce_item(&augmented_grammar) {
-                // Check if this is a reduce by the augmented start rule
-                if let Some(rule) = augmented_grammar
-                    .all_rules()
-                    .find(|r| r.production_id.0 == item.rule_id.0)
-                    && rule.lhs == augmented_start
-                {
-                    if item.lookahead == eof_symbol {
-                        // This is S' -> S • with lookahead $, add accept action
-                        if let Some(&eof_idx) = symbol_to_index.get(&eof_symbol) {
-                            add_action_with_conflict(
-                                &mut action_table,
-                                &mut conflicts_by_state,
-                                state_idx,
-                                eof_idx,
-                                Action::Accept,
-                            );
-                        }
-                    }
-                    // NEVER add a regular reduce action for the augmented start rule
-                    continue;
-                }
-
-                // Regular reduce action
-                if let Some(&rule_id) = production_to_rule_id.get(&item.rule_id.0) {
-                    let rule = &rules[rule_id as usize];
-                    let is_empty_production = rule.rhs_len == 0;
-
-                    // For empty productions, we need to add reduce actions for all symbols in FOLLOW set
-                    let lookaheads_to_check: Vec<SymbolId> = if is_empty_production {
-                        // Get FOLLOW set for the LHS of this rule
-                        if let Some(follow_set) = first_follow.follow(rule.lhs) {
-                            // Map FOLLOW set symbols to actual parse table symbols.
-                            // This replaces EOF_SENTINEL (SymbolId(0)) with the actual eof_symbol.
-                            follow_set
-                                .ones()
-                                .map(|idx| map_follow_symbol(SymbolId(idx as u16), eof_symbol))
-                                .collect()
-                        } else {
-                            vec![item.lookahead]
-                        }
-                    } else {
-                        vec![item.lookahead]
-                    };
-
-                    for lookahead in lookaheads_to_check {
-                        if let Some(&lookahead_idx) = symbol_to_index.get(&lookahead) {
-                            let new_action = Action::Reduce(RuleId(rule_id));
-
-                            // Always add reduce actions - let conflict resolution handle precedence
-                            add_action_with_conflict(
-                                &mut action_table,
-                                &mut conflicts_by_state,
-                                state_idx,
-                                lookahead_idx,
-                                new_action,
-                            );
-                        }
-                    }
-                }
-            }
-            // Note: Shift actions were already added before this loop
-        }
-    }
-
-    // Shift actions were already added before reduce actions
-
-    // Build precedence tables once
-    let production_count = augmented_grammar.all_rules().count() as u32;
-    // token_count includes EOF (symbol 0 in table) plus all regular tokens.
-    let token_count = (internal_tokens.len() + 1) as u32;
-    let prec_tables = build_prec_tables(
-        &augmented_grammar,
-        &symbol_to_index,
-        token_count,
-        production_count,
-    );
-
-    // Calculate the first non-terminal index
-    // Terminals are: EOF + internal tokens + external tokens.
-    // So first non-terminal is at the terminal boundary.
-    let first_nonterminal_idx = internal_tokens.len() + ext_tokens.len() + 1;
-
-    // Resolve conflicts using precedence
-    for ((state_idx, symbol_idx), _actions) in conflicts_by_state {
-        // Guard rail: validate indices
-        debug_assert!(state_idx < action_table.len(), "state_idx out of bounds");
-        debug_assert!(
-            symbol_idx < action_table[0].len(),
-            "symbol_idx out of bounds"
-        );
-
-        // Only resolve on terminal columns (never on gotos).
-        // Terminals occupy indices [0, first_nonterminal_idx).
-        if symbol_idx >= first_nonterminal_idx {
-            continue; // Skip non-terminal columns
-        }
-
-        let cell = &mut action_table[state_idx][symbol_idx];
-
-        // Guard rail: skip empty cells
-        if cell.is_empty() {
-            continue;
-        }
-
-        // If ACCEPT is present, keep it alone (canonical LR(1) accept)
-        if cell.iter().any(|a| matches!(a, Action::Accept)) {
-            *cell = vec![Action::Accept];
-            continue;
-        }
-
-        // Extract first shift and the set of reduces in the cell
-        let first_shift = cell.iter().find_map(|a| {
-            if let Action::Shift(s) = a {
-                Some(*s)
-            } else {
-                None
-            }
-        });
-        let mut reduces: Vec<u16> = cell
-            .iter()
-            .filter_map(|a| {
-                if let Action::Reduce(pid) = a {
-                    Some(pid.0)
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        // If there are multiple reduces, resolve them first (rule precedence)
-        if reduces.len() > 1 {
-            let winner = reduces[1..].iter().fold(reduces[0], |acc, &r| {
-                decide_reduce_reduce(acc, r, &prec_tables)
-            });
-            reduces.clear();
-            reduces.push(winner);
-            // keep the non-reduce actions (shift/accept) as-is for now
-            cell.retain(|a| {
-                matches!(a, Action::Shift(_)) || matches!(a, Action::Reduce(pid) if pid.0 == winner)
-            });
-        }
-
-        // Now we have at most one reduce and at most one shift
-        if let (Some(s), Some(r)) = (first_shift, reduces.first().copied()) {
-            match decide_with_precedence(symbol_idx, r, &prec_tables) {
-                PrecDecision::PreferShift => *cell = vec![Action::Shift(s)],
-                PrecDecision::PreferReduce => *cell = vec![Action::Reduce(RuleId(r))],
-                PrecDecision::Error => {
-                    // Non-associative at equal precedence: forbid combination at this lookahead.
-                    // For GLR you can either force a parse error here or keep both and let runtime err.
-                    // Common Yacc behavior is to make it a syntax error:
-                    // *cell = vec![Action::Error];  // Uncomment if you want to make it a hard error
-                    // For now, keep both for GLR
-                }
-                PrecDecision::NoInfo => {
-                    // For GLR: when no precedence information is available, keep both actions
-                    // This preserves conflicts for GLR runtime to handle via forking
-                    // Don't resolve the conflict - let GLR handle it at runtime
-                }
-            }
-        }
-    }
-
-    // Add non-terminal goto entries to the goto table
-    for ((from_state, symbol), _to_state) in &collection.goto_table {
-        // Check if this symbol is a non-terminal using the tracking from collection
-        let is_terminal = collection
-            .symbol_is_terminal
-            .get(symbol)
-            .copied()
-            .unwrap_or(*symbol == eof_symbol); // EOF is a terminal
-
-        if !is_terminal && let Some(&symbol_idx) = symbol_to_index.get(symbol) {
-            let state_idx = from_state.0 as usize;
-            if state_idx < goto_table.len() && symbol_idx < goto_table[state_idx].len() {
-                // "DEBUG: Setting goto for state {} non-terminal {} (id={}) -> state {}"
-            }
-        }
-    }
-
-    // Fill goto table from collection's goto_table (kept for compatibility)
-    for ((from_state, symbol), to_state) in &collection.goto_table {
-        let from_idx = from_state.0 as usize;
-        if let Some(&symbol_idx) = symbol_to_index.get(symbol) {
-            goto_table[from_idx][symbol_idx] = *to_state;
-        }
-    }
-
-    // Post-process is no longer needed with proper augmentation
-    // The accept action is added when we see S' -> S • with EOF lookahead
-
-    // But we still need to handle the original grammar's symbol mapping
-    if let Some(_start_symbol) = grammar.start_symbol() {
-        // Find all states and check if they need EOF reduce actions
-        for (state_idx, _item_set) in collection.sets.iter().enumerate() {
-            // Skip this post-processing - handled by augmentation
-            let needs_eof_reduce = false;
-            let reduce_rule_id: Option<RuleId> = None;
-
-            // If we found a reduce item that needs EOF action, ensure it's in the action table
-            if needs_eof_reduce
-                && let Some(rule_id) = reduce_rule_id
-                && let Some(&eof_idx) = symbol_to_index.get(&SymbolId(0))
-            {
-                // Check if EOF action already exists
-                if action_table[state_idx][eof_idx].is_empty() {
-                    action_table[state_idx][eof_idx].push(Action::Reduce(rule_id));
-                }
-            }
-        }
-    }
-
-    // Build symbol metadata
-    let mut symbol_metadata = Vec::new();
-
-    // Add terminal symbols
-    for (symbol_id, token) in &grammar.tokens {
-        symbol_metadata.push(SymbolMetadata {
-            name: token.name.clone(),
-            is_visible: !token.name.starts_with('_'),
-            is_named: !matches!(&token.pattern, TokenPattern::String(_)),
-            is_supertype: false,
-            // Additional fields required by API contracts
-            is_terminal: true,
-            is_extra: grammar.extras.contains(symbol_id),
-            is_fragile: false, // TODO: implement fragile token detection
-            symbol_id: *symbol_id,
-        });
-    }
-
-    // Add non-terminal symbols
-    for symbol_id in grammar.rules.keys() {
-        let is_supertype = grammar.supertypes.contains(symbol_id);
-        symbol_metadata.push(SymbolMetadata {
-            name: format!("rule_{}", symbol_id.0),
-            is_visible: true,
-            is_named: true,
-            is_supertype,
-            // Additional fields required by API contracts
-            is_terminal: false,
-            is_extra: false,   // non-terminals are never extra
-            is_fragile: false, // TODO: implement fragile token detection
-            symbol_id: *symbol_id,
-        });
-    }
-
-    // Add external symbols
-    for external in &grammar.externals {
-        symbol_metadata.push(SymbolMetadata {
-            name: external.name.clone(),
-            is_visible: !external.name.starts_with('_'),
-            is_named: true,
-            is_supertype: false,
-            // Additional fields required by API contracts
-            is_terminal: true,             // external symbols are terminals
-            is_extra: false,               // TODO: check if external symbol is in extras
-            is_fragile: false,             // TODO: implement fragile token detection
-            symbol_id: external.symbol_id, // use external symbol ID
-        });
-    }
-
-    // Compute external scanner states
-    // For each state, determine which external tokens are valid
-    // Now we only track validity - transitions are in the main action table
-    let mut external_scanner_states =
-        vec![vec![false; augmented_grammar.externals.len()]; state_count];
-
-    // Create a mapping from external symbol_id to index
-    let mut external_symbol_to_idx = BTreeMap::new();
-    for (idx, external) in augmented_grammar.externals.iter().enumerate() {
-        external_symbol_to_idx.insert(external.symbol_id, idx);
-    }
-
-    // Determine which external tokens are valid in each state
-    // An external token is valid if there's a shift action for it in that state
-    for state_idx in 0..state_count {
-        for (external_idx, external) in augmented_grammar.externals.iter().enumerate() {
-            // Check if this external has a shift action in this state
-            if let Some(&symbol_idx) = symbol_to_index.get(&external.symbol_id) {
-                // Check if any action in the cell is a shift
-                if action_table[state_idx][symbol_idx]
-                    .iter()
-                    .any(|a| matches!(a, Action::Shift(_)))
-                {
-                    external_scanner_states[state_idx][external_idx] = true;
-                }
-            }
-        }
-    }
-
-    // Build nonterminal_to_index mapping
-    let mut nonterminal_to_index = BTreeMap::new();
-    for (&symbol_id, &idx) in &symbol_to_index {
-        if nonterminal_symbols.contains(&symbol_id) {
-            nonterminal_to_index.insert(symbol_id, idx);
-        }
-    }
-
-    let _rule_count = rules.len();
-
-    // Calculate proper counts for EOF symbol
-    // token_count includes EOF (Symbol 0 in table) + all internal tokens
-    let token_count = internal_tokens.len() + 1;
-    let external_token_count = ext_tokens.len();
-
-    // Normalize action table for deterministic output
-    normalize_action_table(&mut action_table);
-
-    // Build reverse map once, keep the source of truth inside the table.
-    let mut index_to_symbol = vec![SymbolId(u16::MAX); symbol_to_index.len()];
-    for (sym, &idx) in &symbol_to_index {
-        index_to_symbol[idx] = *sym;
-    }
-
-    let mut table = ParseTable {
-        action_table,
-        goto_table,
-        symbol_metadata,
-        state_count,
-        symbol_count,
-        symbol_to_index,
-        index_to_symbol,
-        external_scanner_states,
-        rules,
-        nonterminal_to_index,
-        goto_indexing: GotoIndexing::NonterminalMap, // Will be auto-detected
-        eof_symbol,
-        start_symbol: original_start,
-        grammar: grammar.clone(),
-        initial_state: StateId(0), // Default initial state
-        token_count,
-        external_token_count,
-        lex_modes: vec![
-            LexMode {
-                lex_state: 0,
-                external_lex_state: 0
-            };
-            state_count
-        ],
-        extras: vec![],             // TODO: Get from grammar metadata
-        dynamic_prec_by_rule,       // Now properly populated from grammar rules
-        rule_assoc_by_rule,         // Now properly populated from grammar rules
-        alias_sequences: vec![],    // TODO: Get from grammar
-        field_names: vec![],        // TODO: Get from grammar
-        field_map: BTreeMap::new(), // TODO: Get from grammar
-    };
-
-    // Auto-detect GOTO indexing mode
-    table.detect_goto_indexing();
-
-    Ok(table)
-}
-
 /// Sanity check parse table for correctness
 #[must_use = "validation result must be checked"]
 pub fn sanity_check_tables(pt: &ParseTable) -> Result<(), String> {
@@ -3331,45 +2316,8 @@ pub fn sanity_check_tables(pt: &ParseTable) -> Result<(), String> {
     Ok(())
 }
 
-/// Add an action to the parse table, tracking conflicts
-fn add_action_with_conflict(
-    action_table: &mut Vec<Vec<ActionCell>>,
-    conflicts_by_state: &mut BTreeMap<(usize, usize), Vec<Action>>,
-    state_idx: usize,
-    symbol_idx: usize,
-    new_action: Action,
-) {
-    // Bounds check
-    if state_idx >= action_table.len() || symbol_idx >= action_table[0].len() {
-        panic!(
-            "Index out of bounds in add_action_with_conflict: state_idx={}, symbol_idx={}, table_size={}x{}",
-            state_idx,
-            symbol_idx,
-            action_table.len(),
-            if action_table.is_empty() {
-                0
-            } else {
-                action_table[0].len()
-            }
-        );
-    }
-
-    let current_cell = &mut action_table[state_idx][symbol_idx];
-
-    // Check if this action already exists
-    if !current_cell.iter().any(|a| action_eq(a, &new_action)) {
-        // Add the action to the cell
-        current_cell.push(new_action.clone());
-
-        // If there are now multiple actions, track as a conflict
-        if current_cell.len() > 1 {
-            let entry = conflicts_by_state
-                .entry((state_idx, symbol_idx))
-                .or_default();
-            *entry = current_cell.clone();
-        }
-    }
-}
+#[cfg(test)]
+use automaton::actions::normalize_action;
 
 /// Build LR(1) automaton using the GlrResult type alias
 ///
@@ -3380,20 +2328,6 @@ pub fn build_lr1_automaton_res(
     first_follow: &FirstFollowSets,
 ) -> GlrResult<ParseTable> {
     build_lr1_automaton(grammar, first_follow)
-}
-
-/// Check if two actions are equivalent
-fn action_eq(a: &Action, b: &Action) -> bool {
-    match (a, b) {
-        (Action::Shift(s1), Action::Shift(s2)) => s1 == s2,
-        (Action::Reduce(r1), Action::Reduce(r2)) => r1 == r2,
-        (Action::Accept, Action::Accept) => true,
-        (Action::Error, Action::Error) => true,
-        (Action::Fork(a1), Action::Fork(a2)) => {
-            a1.len() == a2.len() && a1.iter().zip(a2).all(|(x, y)| action_eq(x, y))
-        }
-        _ => false,
-    }
 }
 
 #[cfg(test)]
